API methods returning `Iterable` breaks RESTclient
It is impossible to use the official rest-client module on API methods
that return `Iterable`, as Jackson does not know which concrete type to
implement. The Jackson configuration is hidden behind RESTeasy, so
reconfiguring Jackson is non trivial.

This PR changes some API calls from `Iterable` to `List`, which
RESTeasy/Jackson is able to deserialize. Further PRs will need to follow
for other `Iterable`-returning API methods.