Manual URL manipulation leads to invalid results.
The test was working by accident on UNIX - with an absolute path the
beginning of the string would be file:///, which is the expected prefix.
On Windows though the result was an invalid URL. Use the platform
provided functionality to construct the URL and remove invalid
statement in the comment.