Load URL referenced YAMLs.
Allow URL references to YAML files in addition to catalog IDs and java types.
Currently limited to services (PlatformComponentTemplate level).

In the case of more than one service in the referenced plan just use the wrapped app as a child.
The template referencing the yaml will override the name, IDs of the child. Locations, policies,
effectors, ect. will be appended to the child's config. Do the same for catalog item references.