Fix rebinding to entities/policies living in OSGi bundles
When an entity/policy is created from a catalog item we make a note of its related catalog item id. On rebind we use that information to get the class loading context from the corresponding catalog item so that classes from OSGi bundles referenced in the item config are accessible.