Make embedded wars optional
Modifies code that relies on war files as an embedded resource to
gracefully handle cases where the files are missing. Limited to tests,
which will cause TestNG to skip the test if the resource is missing.
Required for our source distribution, as Apache convention is to not
ship binaries in releases.

Conflicts:
	software/webapp/src/test/java/brooklyn/entity/webapp/ElasticJavaWebAppServiceIntegrationTest.java
	software/webapp/src/test/java/brooklyn/entity/webapp/jboss/Jboss6ServerEc2LiveTest.java
	software/webapp/src/test/java/brooklyn/entity/webapp/jboss/Jboss7ServerEc2LiveTest.java
	software/webapp/src/test/java/brooklyn/entity/webapp/jboss/Jboss7ServerIntegrationTest.java