Logging brooklyn-persister thread Jclouds messages in a separate file
When a server with persistence to an object store enabled manages lots of
entities the Jclouds log is overwhelmed by messages logged from the
brooklyn-persister thread (in one case logging at a rate of ~ half a
million lines every four minutes).

The appender added in this commit alleviates - a little! - this logging
overload.

Anecdotal measurements of the ratio between the persister log and jclouds
log sizes, with EmptySoftwareProcess entities:
0 entities: persister log 0.01 times size of jclouds log
1 entity: 0.36
5 entities: ~1.13
10 entities: 1.75
50 entities: 8.6