ResourcePredicates.urlExists()
Adds a predicate that is true if a resource is available via a
configurable context object. ConfigConstraints looks for instances of the
Brooklyn-object-aware predicate and calls apply(T, BrooklynObject) if
possible.

Requires non-deployment management contexts to return useful entitlement
and OSGi managers.

Rework contextual predicates

Context must be passed to apply, not set as a field on the predicate.